+++ CHANGES.rst
@@ -234,6 +234,10 @@ DAL node
 Protocol
 ~~~~~~~~
 
+- Add a new RPC ``GET
+  /chains/<chain>/blocks/<block>/helpers/total_baking_power`` to retrieve the
+  total baking power for the current cycle. (MR :gl:`!17553`)
+
 Miscellaneous
 -------------